+++ FE/css/ProfilePage/matchLog.css
@@ -1,0 +1,225 @@
+.matchLog{
+    width: 100%;
+    height: 100%;
+    display: flex;
+	flex-direction: row;
+	justify-content: space-around;
+	align-items: center;
+	position: relative;
+}
+
+.matchDate{
+    position: absolute;
+    top: 5%;
+    left: 3%;
+    font-size: 1.1rem;
+    font-weight: 200;
+    color: white;
+    opacity: 60%;
+}
+
+.matchScoreContainer{
+    width: 12%;
+    height: 60%;
+    display: flex;
+    margin-top: 1%;
+	flex-direction: column;
+	justify-content: space-between;
+	align-items: center;
+}
+
+
+.matchResultImage{
+    width: 100%;
+    height: 50%;
+    overflow: hidden;
+}
+
+.scoreFrame{
+    height: 50%; 
+    width: 100%;
+    display: flex;
+	flex-direction: row;
+	justify-content: center;
+	align-items: center;
+    font-size: 2rem;
+    font-weight: 200;
+}
+
+.matchTeamListContainer{
+    width: 30%;
+    height: 40%;
+    display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.matchTeamContainer{
+    width: 35%;
+    height: 100%;
+    display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.matchVampireTeamImage{
+    border-radius: 50%;
+    height: 100%;
+    aspect-ratio: 1;
+    overflow: hidden;
+    /* background: url('../../images/vampireIcon.png') center/cover; */
+    filter: drop-shadow(0px 0px 20px #ff2500CC);
+}
+
+.matchHumanTeamImage{
+    border-radius: 50%;
+    height: 100%;
+    aspect-ratio: 1;
+    overflow: hidden;
+    /* background: url('../../images/vampireIcon.png') center/cover; */
+    filter: drop-shadow(0px 0px 20px #4092f6CC);
+}
+
+.vampireAbilityImage{
+    border: 0.06rem solid white;
+    width: 25%;
+    aspect-ratio: 1;
+    border-radius: 0.2rem;
+    overflow: hidden;
+    align-self: flex-end;
+    margin-bottom: 7%;
+}
+
+.vsFrame{
+    width: 30%;
+    display: flex;
+	flex-direction: row;
+	justify-content: center;
+	align-items: center;
+    font-size: 2rem;
+    font-weight: 100;
+}
+
+.matchPlayerListContainer{
+    width: 30%;
+    height: 100%;
+    display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.teamPlayerListContainer{
+    width: 45%;
+    height: 100%;
+    display: flex;
+	flex-direction: column;
+    justify-content: space-evenly;
+	align-items: center;
+}
+
+.playerContainer{
+    width: 100%;
+    height: 15%;
+    box-sizing: border-box;
+    display: flex;
+	flex-direction: row;
+    justify-content: space-between;
+	align-items: center;
+}
+
+.playerAvatarImgFrame{
+	height: 100%;
+	aspect-ratio: 1/1;
+	overflow: hidden;
+    background: linear-gradient(to bottom, #291350 0%, #02041F 100%);
+}
+
+.playerAvatarImg{
+    width: 100%;
+    height: 100%;
+}
+
+.noAvatarRed{
+    width: 100%;
+    height: 100%;
+    position: relative;
+}
+
+.noAvatarRed::before, .noAvatarRed::after {
+    content: '';
+    position: absolute;
+    top: 50%; 
+    left: 50%; 
+    width: 240%; 
+    height: 0.01rem;
+    background-color: #ff2500;
+}
+
+
+.noAvatarRed::before {
+    transform: translate(-50%, -50%) rotate(45deg); 
+}
+
+.noAvatarRed::after {
+    transform: translate(-50%, -50%) rotate(-45deg);
+}
+
+.noAvatarBlue{
+    width: 100%;
+    height: 100%;
+    position: relative;
+}
+
+
+.noAvatarBlue::before, .noAvatarBlue::after {
+    content: '';
+    position: absolute;
+    top: 50%; 
+    left: 50%; 
+    width: 240%; 
+    height: 0.01rem;
+    background-color: #4092f6;
+}
+
+
+.noAvatarBlue::before {
+    transform: translate(-50%, -50%) rotate(45deg); 
+}
+
+.noAvatarBlue::after {
+    transform: translate(-50%, -50%) rotate(-45deg);
+}
+
+.redBorder{
+    border: 0.05rem solid #ff2500CC;
+    border-radius: 0.3rem;
+}
+
+.blueBorder{
+    border: 0.05rem solid #4092f6CC;
+    border-radius: 0.3rem;
+}
+
+
+
+
+.playerNickname {
+    width: 70%;
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    overflow: hidden; /* 오버플로우 설정 */
+}
+
+.playerNickname span {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    flex-shrink: 1; /* 기본값은 1이므로 아이템이 줄어들 수 있게 설정 */
+    min-width: 0; /* 축소할 수 있는 최소 너비를 설정하여 ellipsis가 동작하도록 함 */
+}
