+++ FE/js/ProfilePage/ProfilePageManager.js
@@ -1,4 +1,5 @@
 import { GlobalConnectionError, ProfileTargetNotFound, isSocketConnected } from "../Error/Error.js";
+import { getMatchLogDiv } from "./getMatchLogDiv.js";
 
 class ProfilePageManager {
 	constructor(app, clientInfo, renderPage, queryParam) {
@@ -102,85 +103,10 @@ class ProfilePageManager {
                 </button>
             </div>
             <div id="matchLogContainer">
-                ${this.profileTarget.gameHistoryList.map(gameHistory => this._getMatchLogDiv(gameHistory)).join("")}
+                ${this.profileTarget.gameHistoryList.map(gameHistory => `<div class="matchLogPanel">${getMatchLogDiv(gameHistory)}</div>`).join("")}
             </div>
         </div>
         `;
-	}
-
-	_getMatchLogDiv(gameHistory) {
-		const date = new Date(gameHistory.timestamp);
-		const year = date.getFullYear();
-		const month = String(date.getMonth() + 1).padStart(2, "0");
-		const day = String(date.getDate()).padStart(2, "0");
-		return `
-        <div class="matchLog" data-id="${gameHistory.gameId}">
-            <div class="matchDate">
-                ${year}.${month}.${day}
-            </div>
-            <div class="matchScoreContainer">
-                <img class="matchResultImage" src="images/${gameHistory.result?.toLowerCase()}Icon.svg">
-                <div class="scoreFrame">
-                    ${gameHistory.score[0]} : ${gameHistory.score[1]}
-                </div>
-            </div>
-            <div class="matchTeamListContainer">
-                <div class="matchTeamContainer">
-                    ${gameHistory.ability[0] !== "none" ? "<img class='vampireAbilityImage' src='images/ability/" + gameHistory.ability[0] + ".png'>" : "<div></div>"}
-                    <img class="match${gameHistory.teamKind[0] === "human" ? "Human" : "Vampire"}TeamImage" src="images/${gameHistory.teamKind[0]}Icon.png">
-                </div>
-                <div class="vsFrame">VS</div>
-                <div class="matchTeamContainer">
-                    <img class="match${gameHistory.teamKind[1] === "human" ? "Human" : "Vampire"}TeamImage" src="images/${gameHistory.teamKind[1]}Icon.png">
-                    ${gameHistory.ability[1] !== "none" ? "<img class='vampireAbilityImage' src='images/ability/" + gameHistory.ability[1] + ".png'>" : "<div></div>"}
-                </div>
-            </div>
-            ${this.getMatchPlayerListContainerDiv(gameHistory.myTeamClientInfoList, gameHistory.opponentTeamClientInfoList)}
-        </div>       
-        `;
-	}
-
-	getMatchPlayerListContainerDiv(myTeamClientInfoList, opponentTeamClientInfoList) {
-		return `
-        <div class="matchPlayerListContainer">
-            <div class="teamPlayerListContainer">
-                ${myTeamClientInfoList.map(player => this.getPlayerContainerDiv(player, "red")).join("")}
-                ${Array.from({ length: 5 - myTeamClientInfoList.length })
-									.map(() => this.getEmptyPlayerContainerDiv("red"))
-									.join("")}
-            </div>
-            <div class="teamPlayerListContainer">
-                ${opponentTeamClientInfoList.map(player => this.getPlayerContainerDiv(player, "blue")).join("")}
-                ${Array.from({ length: 5 - opponentTeamClientInfoList.length })
-									.map(() => this.getEmptyPlayerContainerDiv("blue"))
-									.join("")}
-            </div>
-        </div>
-        `;
-	}
-
-	getPlayerContainerDiv(player, color) {
-		return `
-        <div class="playerContainer" data-id="${player.id}">
-            <div class="playerAvatarImgFrame ${color}Border">
-                <img class="playerAvatarImg" src="${player.avatarUrl}">
-            </div>
-            <div class="playerNickname">
-                <span>${player.nickname}</span>
-            </div>
-        </div>
-        `;
-	}
-	getEmptyPlayerContainerDiv(color) {
-		return `
-        <div class="playerContainer">
-             <div class="playerAvatarImgFrame ${color}Border">
-                 <div class="noAvatar${color === "red" ? "Red" : "Blue"}"></div>
-             </div>
-             <div class="playerNickname">
-                 <span></span>
-             </div>
-         </div>`;
 	}
 }
